Initial parser implementation
This implements a very simple parser that just reads rules, roles and
actions with one element. Given that go doesn't read the given YAML or
JSON in an ordered manner, we test that each entry should be in the
output instead of testing the whole output itself (this way we don't
care about the order).

This also contains a small mock main.go which one can execute, this will
eventually be replaced by an actual CLI to parse policy.json files and
output rego policies.